✨ feat(acceptance): show interaction cost to the reviewer, not just the owner
Two fixes for the same root problem — the KLM cost existed but nobody who needed
it could see it.

**The scaffolder wrote the null.** `report-init.sh` emitted
`"interactionCost": null` to document the field, which is what made the previous
commit's ingest guard misfire. Remove it at the source: a scaffolded report now
simply omits the key, and ingest prices its trace. The guard stays as defence in
depth for hand-written reports, but the scaffolder no longer arms it.

**The panel was in the wrong place.** `InteractionCostPanel` lived only in the
round report, which opens from owner-scoped round history
(`canViewAcceptanceHistory = isOwner`). So the one audience the number exists for
— a reviewer weighing whether a flow is worth accepting — could never see it,
while the owner who already knows the cost could. Surface it on the acceptance
page itself, collapsed by default, showing total/active/wait inline and the full
per-phase breakdown on expand. No server or permission change is needed: the cost
already ships in the shared bundle (only `origin` is stripped for visitors).

Phases attributed to a check (`--klm-check` → `phase.checkItemId`) now render
that check's title instead of an internal slug, so the cost reads against the
thing being judged.

Splits `ReportViewer` (1956 → 1537 lines) by extracting the panel, and its pure
pricing/reading helpers into `interactionCostModel.ts` so round selection is
unit-testable without mounting React — the repo does not take new component
tests.
