+++ .agents/acceptance/scripts/report-init.sh
@@ -125,7 +125,6 @@ cat > "$DIR/result.json" << EOF
   "entry": "",
   "plan": [],
   "cases": [],
-  "interactionCost": null,
   "summary": { "total": 0, "passed": 0, "failed": 0, "blocked": 0, "verdict": "pending", "conclusion": "", "score": null }
 }
 EOF
